loaders: added support for the "file://" scheme in loaders
"file://" is resolved with the local loader
