+++ yaml_test.go
@@ -17,7 +17,6 @@ package swag
 import (
 	"encoding/json"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -32,29 +31,6 @@ func (f failJSONMarshal) MarshalJSON() ([]byte, error) {
 	return nil, errors.New("expected")
 }
 */
-
-func TestLoadHTTPBytes(t *testing.T) {
-	_, err := LoadFromFileOrHTTP("httx://12394:abd")
-	assert.Error(t, err)
-
-	serv := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
-		rw.WriteHeader(http.StatusNotFound)
-	}))
-	defer serv.Close()
-
-	_, err = LoadFromFileOrHTTP(serv.URL)
-	assert.Error(t, err)
-
-	ts2 := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
-		rw.WriteHeader(http.StatusOK)
-		_, _ = rw.Write([]byte("the content"))
-	}))
-	defer ts2.Close()
-
-	d, err := LoadFromFileOrHTTP(ts2.URL)
-	assert.NoError(t, err)
-	assert.Equal(t, []byte("the content"), d)
-}
 
 func TestYAMLToJSON(t *testing.T) {
 
@@ -117,35 +93,6 @@ name: a string value
 	d, err = YAMLToJSON(dd)
 	assert.NoError(t, err)
 	assert.Equal(t, json.RawMessage(`{"description":"object created"}`), d)
-}
-
-func TestLoadStrategy(t *testing.T) {
-
-	loader := func(p string) ([]byte, error) {
-		return []byte(yamlPetStore), nil
-	}
-	remLoader := func(p string) ([]byte, error) {
-		return []byte("not it"), nil
-	}
-
-	ld := LoadStrategy("blah", loader, remLoader)
-	b, _ := ld("")
-	assert.Equal(t, []byte(yamlPetStore), b)
-
-	serv := httptest.NewServer(http.HandlerFunc(yamlPestoreServer))
-	defer serv.Close()
-
-	s, err := YAMLDoc(serv.URL)
-	assert.NoError(t, err)
-	assert.NotNil(t, s)
-
-	ts2 := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
-		rw.WriteHeader(http.StatusNotFound)
-		_, _ = rw.Write([]byte("\n"))
-	}))
-	defer ts2.Close()
-	_, err = YAMLDoc(ts2.URL)
-	assert.Error(t, err)
 }
 
 var yamlPestoreServer = func(rw http.ResponseWriter, r *http.Request) {
